refactor: refactor project structure
