+++ src/app/app.component.html
@@ -1,1 +1,7 @@
-<app-podcast-page></app-podcast-page>
+<app-header>
+<!--    (searchClicked)="initListLoad()"-->
+<!--    (sort)="onSort($event)"-->
+<!--    (search)="onSearch($event)"-->
+<!--&gt;-->
+</app-header>
+<router-outlet></router-outlet>
