+++ src/app/core /components/header/header.component.scss
