+++ src/app/youtube/components/main-page/main-page.component.html
@@ -1,0 +1,6 @@
+<div class="podcast-page">
+  <app-podcast-list
+    [podcasts]="podcastsArr | filter:search"
+  ></app-podcast-list>
+</div>
+x
