+++ src/app/youtube/components/main-page/podcast-list/podcast-item/podcast-item.component.scss
