+++ src/app/youtube/components/main-page/podcast-list/podcast-list.component.scss
