+++ src/app/youtube/components/main-page/podcast-list/podcast-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { PodcastItem } from '../../../core /models/podcast-item.typing';
+import { PodcastItem } from '../../../models/podcast-item.typing';
 
 @Component({
   selector: 'app-podcast-list',
