+++ src/app/youtube/models/podcast-item.typing.ts
