+++ src/app/youtube/youtube-routing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { MainPageComponent } from './components/main-page/main-page.component';
+
+const routes: Routes = [
+    { path: '', component: MainPageComponent },
+];
+
+@NgModule({
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule],
+})
+export class YoutubeRoutingModule {
+}
